Refactor DataTable to be a reusable component
